updated
everything changed dawg

+++ app/controllers/comments_controller.rb
@@ -1,4 +1,6 @@
 class CommentsController < ApplicationController
+  before_action :authorize_user, only: [:edit, :update, :destroy]
+
   def create
     @comment   = Comment.new comment_params
     @article = Article.find params[:article_id]
@@ -12,6 +14,7 @@ class CommentsController < ApplicationController
     def destroy
       comment = Comment.find params[:id]
       article = Article.find params[:question_id]
+      redirect_to root_path, alert: "access denined" unless can? :destroy, @comment
       comment.destroy
       redirect_to article_path(article), notice: "Comment deleted"
     end
